v0.6a - Recording Prototyping
 -- UI --
Moved over from a single UI class to a master class that is added to the player with different states.
Moved all recording assets to InGame_Default and added an InGame_Default to the master class' widget switcher

 -- Controllers --
Changed MoveX to have an argument of FInputActionInstance instead of FInputActionValue
 - This allows access to more data about the input, such as elapsed time

 -- Character --
Added CharacterDefault, a subclass of Character_Parent
 - To be used instead of the Character_Parent as a default class
Moved MenuUI setup from Tick to BeginPlay
 - This wasn't supposed to be in Tick in the first place
Made MoveX, MoveY and Interact inputs virtual, so they could be overriden.
Added a float TimerTickRate, used by recording and playback characters

-- Recording --
Originally, I was going to simply use Tick with it's DetlaTime output to record the direction of the Recording character on each frame.  However, if the FPS would fluctuate the character would be recorded wrong.  For example, if the recording was captured at 60fps but played back at 30, the character would be much slower in the playback.

This is avoided via the use of a timer with a set loop rate.  Both the recording and playback will use a timer with the same 'tick' rate to maintain the speed of the character.

Added Character_Record, used to record a ghost for playback
Added struct FRecordingData, used to store the value of an input at a tick
 - Currently in Character_Record but needs to be moved as the playback character will use this struct too
In the constructor, Character_Record finds and stores the input context data.  This data is used to setup inputs specific to this character, which binds to the inputs Started and Complete state
 - When an input is started, the respective boolean is set to true.  When the input is complete, the boolean is set to false.
Setup and starts two timers, one for the recordings total time and one for the recording tick.
 - Once the recording reaches the total time, the recording ends
 - Each tick each input bool is checked to see if they are true.  If they are, a new index is added to the respective TArray with the input's value and the current tick time