v0.7a - Playback Character
 -- UI --
OnRecordButtonReleased now finds the LevelController in the world and calls StartLevelPlayback instead of calling StartRecording

-- Mechanics - Record Pad --
Added PlaybackCharacterClass property, used by the BluePrint class to set what ACharacter_Playback should be used for playback
Added a pointer to the CurrentCharacter
Implemented StartPlayback function, used to spawn and start a PlaybackCharacter
 - First checks if a record is present.
 - If so, then spawns a Playback character of the set class and calls SetupPlayback on the character, passing through the recording
Also implemented EndPlayback, which simply checks if there is a CurrentCharacter set.  If so, the character is destroyed.

-- Level - LevelController --
Removed UpdatePlayerUIElements, as this is done via the UI class itself
Added a TArray of all RecordPads
Implemented StartLevelPlayback
 - Simply compares all pointers in the RecordPad array with the inputted parameter.  If they match, then StartRecording is called on that RecordPad.  Else, StartPlayback is called instead.
Also implemented EndLevelPlayback
 - Calls EndPlayback on all RecordPads in the array

 -- Character - Record --
Changed MaximumRecordingTime from 5 seconds to 60 seconds
Updated recording camera inputs from using the MoveYBind to the CameraBind
 - This also didn't provided the intended result, so now stores the control rotation of the character
EndRecording now finds the LevelController and calls EndLevelPlayback

 -- Character - Parent --
Changed TimerTickRate from 0.1 to 0.005
Updated RotateCamera to a virtual function, allowing it to be overridden
Updated the interact debug line to use the same location start as the interact trace

-- Character - Playback --
Implemented class
BeginPlay simply casts to the controller and stores it
Overrides MoveX
 - Works the same as the Parent class, but instead uses the rotation of the FirstPersonCamera instead of the control rotation to calculate the direction
Overrides RotateCamera
 - Instead of changing the control rotation, updates the FirstPersonCamera's Pitch and Yaw and the Actor's Yaw
 - PhysicsHandle location is updated the same way as Parent
Implemented SetupPlayback
 - First, stores the playback from the parameter
 - Next, calculates the largest tick value from all the input recordings
 -- Originally tried to use a c++ array and use std::max_element, but this didn't work
 -- Instead, gets the end index tick for each input, stores them in a TArray and iterates over the array for the largest value
 - Finally, sets two timers, one for the total duration and one for the tick
Implemented EndPlayback
 - Simply clears the tick and total timers
Implemented PlaybackTick
 - For each input recorded, checks the tick value at the 0 index of an array.  If it is greater than or equal to the current tick time, calls the corresponding input and removes the index from the array

 -- Characters --
Created BP classes for the Playback Character
Updated the locations of the Record Character's components and set a test mesh