v0.7b - WorldObjects
 -- Mechanics - WorldObjects --
Implemented class
Added getter and setter functions for CurrentGrabber
Updated PROTO_WorldObject to use this class

 -- Characters - Parent --
Updated Interact to check if the object has the WorldObject class.  If so, checks if the object is currently being grabbed.  If so, the object is ungrabbed and then regrabbed by the new character