v0.8c - Level Spawning Part Two
 -- Mechanics - Record Pad --
Fixed a spawning bug by increasing the SetActorLocation to 200 on the Z axis

 -- Mechanics - Parent --
Changed ObjectName property type from a FSting to FName

 -- Levels - LevelDoor --
Added a Enum property to denote each door

 -- Level - LevelController --
Added a pointer to both the EntranceDoor and ExitDoor
Added a FVector property used to offset the level
Added a TArray to all mechanics in the level
Added a RecordClass pointer, used to set the class to spawn record pads
Removed the Room_Parent pointer
Added a StaticMeshComponent LevelMesh, used to display the mesh of the level
Moved Door setup from RoomParent to LevelController.  Also now setup on BeginPlay
Updated SetupLevel
 - No longer spawns a RoomParent class in the world
 - Now updates the entrance and exit portal locations, followed by spawning in the required RecordPads
 - Followed with the Mechanic Objects and setup their outputs where necessary
 - Finally, the level mesh is set
Changed StartLevelPlayback and EndLevelPlayback to no longer use the ActiveRoom, instead using the RecordPad TArray in the controller
Also updated SetDoorState on both functions to use the EntranceDoor in the controller

This second attempt worked much better than the first attempt, but takes a slightly different path of storing the level data.  Instead of using a single class per level, a row in the data table now stores all of the mechanic details, the record pad locations and the portal door locations in a struct.  This also allowed me to create a Editor Utility tool to automatically take a level that I have created in the editor and setup a new data table row with a click of a widget button.

 -- Level - LevelData --
Added FRecordPadData, which stores the Transform and Index of RecordPads in a level
Added FMechanicData, which stores the ID, Transform, Class, OutputIDs, TriggerRequirement and DefaultObjectState of each mechanic in a level
Added LevelMesh, EntranceDoorTransform, ExitDoorTransform, RecordPad TArray and Mechanics TArray to LevelData

This was the second time I created a Editor Utility Blueprint and I feel like this works well to what it needs.  It successfully transfers the data from the level I create into a single data struct which can then be used to re-spawn the level in game.  It also allows me to bypass the issue that came up in attempt one - where pointers could not be set between child actors in a class

 -- OutLevelMeshes --
Added different LevelMeshes created via the CreateNewLevel tool

 -- Maps --
Added LevelCreationMap, a blank map used to create the levels then convert them into FLevelData

 -- LevelCreationTool --
Added a Editor Utility Widget, used to access the CreateNewLevel tool
 - First, takes in a FName to set the level too via the widget (see Images/v8-004)
 - Next, spawns the GetLevelData actor into the world
 - Next, checks if the InputText is empty.  If so, set OutText to the default value (Images/v8-005)
 - Following this, gets all StaticMeshActors in the level and Merges them into a new MergedStaticMesh with the name inputted (Images/v8-006)
 - With the new merged mesh created, call StartGeneratingLevelData on the GetLevelData actor, passing through the LevelName and the mesh
 - In StartGeneratingLevelData, start by finding all of the level doors and storing the Transform of the EntranceDoor and ExitDoor (Images/v8-007)
 - Next, do the same with the RecordPads but also store their index (Images/v8-008)
 - Follow with the MechanicObjects, storing all necessary information required by MechanicData (Images/v8-009)
 - Finally, make FLevelData with the LevelName, mesh and the data collected and output it (Images/v8-010)
 - With this output, add a new DataTable row to the DT_LevelData and after a delay, destroy the GetLevelData actor (Images.v8-011)
 -- The Data Tabel will require saving to keep the new row(s)