v0.8f - Removing Recording Character
 -- Characters --
Modified and updated code for all four character classes
 - Updated Parent class to only implement functions/properties required by both classes
 -- Removed ToggleMenu, UIInteract, MenuWidgetComponent, WidgetInteractionComponent and MenuUI pointer, as these were only required by the Default Character
 - Added all pieces removed in the Parent class to the Default class
 - Combined the Record class and Default class together into one
 -- This required a slight rework of the Controller_Parent class and the RecordPad class
 -- Items added to the Default class from other classes include:
 --- Functions: ToggleMenu, UIInteract, RecordMoveX, RecordMoveY, RecordCamera, StartRecording, EndRecording, RecordingTick
 --- Components: MenuWidgetComponent, WidgetInteractionComponent, Input Value Binds, Input Active Bools, Recording Timer Handles, Max and Current Tick times, TArrays of Recorded Inputs and MenuUI pointer

I merged the Default class and Record class together as this would allows me to finish levels easier.  Before, when a level had finished and the player had used the ExitDoor, they would need to re-possess the Default character, which then needed to teleport to the same location as the current Record character.  This would require the LevelDoor class to find the Default character in the world, teleport it to an exact location and then call RePossess to keep the illusion of the portal.

 -- UI - RecordingUi --
Changed Cast from Character_Record to Character_Default

 -- Controller --
Added Recording Inputs for RecordMoveX, RecordMoveY and RecordCamera
Changed the pointer type of Character from Character_Parent to Character_Default
Changed the accessability of ToggleMenu from protected to public, so the Character can call to update their menu when required
Updated OnPossess to no longer check if the class is of Character_Record or Character_Default
 - Also binds the action values to the Character_Default's binds on possession
Changed inputs using GetActiveCharacter to just use Character

 -- RecordPad --
Start and SetRecording now have an argument of Character_Default instead of Controller
StartRecording no longer spawns a Character_Record, rather now calls StartRecording on the Character_Default
SetRecording no longer calls the controller to RePossess, instead moves the character to the pad's location