v0.8g - Standing Button and WorldObjectSpawner
 -- Mechanics - Inputs --
Added ModifyVisualElements in the default ToggleInputActive
Mode ToggleInputActive BlueprintCallable

 -- Mechanics - Outputs --
Added ModifyVisualElements in the default ToggleOutput

 -- Mechanics - StandingButton --
Added a input button the player can interact with by using their Interact key
Sets gameplay tag IsInteractable in constructor
Added function UseButton, called by the player to use the button
On UseButton, simply enables the button then disables the button after a delay (Images/v8-012)

-- Mechanics - WorldObjectSpawner --
Added an output object that can spawn WorldObjects when activated
Added a component to control the ObjectSpawnLocation
Added a pointer to the spawned object and a class pointer to control what world object is spawned
On ToggleOutput, the input argument is checked to be true as we only want the box to move/spawn on activation.  Next, the pointer is checked if it is set (an object is spawned).  If one has, it is moved to the ObjectSpawnLocation.  If one hasn't, it is spawned now.

 -- Character - Parent --
Updated Interact to check if the hit actor has an IsInteractable tag.  If so, cast to the SButton and call UseButton
 - The use of a gameplay tag allows more inputs to be interacted with later on in development (if required)