v0.8h - End Record on Room Complete
 -- RecordPad --
Added a default property for the PlayerCharacter argument in SetRecording
Now only teleport the character back to the pad if PlayerCharacter is valid

 -- Mechanics - WOSpawner --
Fixed incorrect comparison, forgot extra '='

 -- LevelDoor --
On teleport, checks if the character is a Default character.  If it is, then cast to the Character_Default class.
After casting, check if the character is currently recording.  If they are, end their recording.

 -- Characters - Default --
Moved bCurrentlyRecording from protected to public
Added a boolean argument to EndRecording with a default property of true
Now checks if the character should be teleported back to their record pad.  If true, their menu is also toggled
Removed the menu toggle from the end of EndRecording