v0.8i - Level Select
 -- RecordPad --
Fixed a playback bug by moving the player to the record pad on recording start

 -- MechanicObject - Parent --
Added a ObjectToSpawn property, used by certain Output mechanics to spawn WorldObjects

 -- MechanicObject - WorldObjectStart --
Swapped the parent class from MechanicObject_Parent to MechanicObject_Output

 -- MechanicObject - WorldObjectSpawner --
Now overrides ResetToDefault from MechanicObject_Parent
 - Simply moves the object under the map

-- LevelData --
Added a ObjectToSpawn property

 -- SelectorData --
Added FPlanetData struct, used to store data about a planet in the LevelSelector
Added FSectorStruct, used to store a collection of planets together
Added FPlanetVisualsData, used to store the components of each planet spawned in the LevelSelector

 -- LevelDoor --
Moved UpdateClipPlane from protected to public

 -- LevelController --
No longer calls SetupLevel on BeginPlay
SetupLevel now clears the level before searching the data table
Also updates the clip plane of the Entrance and Exit doors and sets ObjectToSpawn on the OutputMechanics
Implemented ClearLevel, which simply destroys all objects in each mechanic array (excluding the ExitDoor) and then empties the arrays
Updated StartLevelPlayback from "AMechanicObject_Parent* j : ActiveRoom->GetMechanicObjects()" to "AMechanicObject_Output* j : OutputMechanics"

 -- LevelSelector --
Implemented class, added a camera used to blend the player's view target between
On BeginPlay, find and store a pointer to the LevelController
Implemented PrimaryInteract, which fires a trace from the player's mouse position in the mouse direction.  The trace is then checked to see if it hits a component.  If it does, it queries the PlanetVisuals array to find a matching component, then calls SetupLevel based on the component found.
Implemented SetActiveSector, which finds the data table row of the inputted parameter.  If one is found, then spawns component sets to match the amount of PlanetIDs in the sector via recursion, then updates the planet sets to match the visuals in the found row (Images/v8-014)
Implemented SetActive and EndInteraction, which simply set/clear a pointer to the player interaction with the selector
Added AddNewPlanetVisual, which spawns a spring arm component, static mesh component and rotating movement component set until the target is reached

 -- Characters - Default --
Added a pointer to the Controller_Player to avoid multiple casts
Added a pointer to a LevelSelector, used when interacting with one
Implemented ExitInteraction, which resets the player's view to their FirstPersonCamera
Overrided Interact, now has an extra query on the InteractTrace
 - Checks if the object is a LevellSelector.  If it is, it is casted to, the player's view target is blended with the selectors camera and the CurrentLevelSelector is set to active
 - Additionally, if the player is currently using a level selector, interacting again will exit

 -- Levels --
Removed unneeded merged meshes
Created three tutorial levels
Added an output message to the CreateNewLevel (Images/v8-0015 + 016)